Refactor Names Assignees: @SantiRey Labels: in progress Fixes #4
